tests/mediacapture-depth/BAK/GUM-trivial-constraint-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Trivial mandatory constraint in getUserMedia</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#idl-def-mediatracksettings">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Description</h2>
<p>
  This test checks that setting a trivial mandatory
constraint (width &gt;=0) in getUserMedia works.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>

<script>

promise_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}, width: {min:0}}})
  .then(stream => {
    assert_equals(stream.getVideoTracks().length, 1, "the media stream has exactly one video track");
  })
  .catch(error => {
    assert_unreached("a Video stream of minimally zero width can always be created");
  });
}, "Tests that setting a trivial mandatory constraint in getUserMedia works");

</script>
